Move alias supporting code to locator.
